# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Memory Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FolderLibraryType =
  (typeof FolderLibraryType)[keyof typeof FolderLibraryType];

export const FolderLibraryType = {
  memory: "memory",
  plan: "plan",
  material: "material",
} as const;

export interface Folder {
  id: number;
  name: string;
  parentId?: number | null;
  libraryType: FolderLibraryType;
  createdAt: string;
  children?: Folder[];
  memoryCount?: number;
}

export type CreateFolderRequestLibraryType =
  (typeof CreateFolderRequestLibraryType)[keyof typeof CreateFolderRequestLibraryType];

export const CreateFolderRequestLibraryType = {
  memory: "memory",
  plan: "plan",
  material: "material",
} as const;

export interface CreateFolderRequest {
  name: string;
  parentId?: number | null;
  libraryType: CreateFolderRequestLibraryType;
}

export interface UpdateFolderRequest {
  name?: string;
  parentId?: number | null;
}

export type MemoryCardEmotionLabel =
  (typeof MemoryCardEmotionLabel)[keyof typeof MemoryCardEmotionLabel];

export const MemoryCardEmotionLabel = {
  calm: "calm",
  excited: "excited",
  positive: "positive",
  negative: "negative",
  neutral: "neutral",
} as const;

export type MemoryCardLibraryType =
  (typeof MemoryCardLibraryType)[keyof typeof MemoryCardLibraryType];

export const MemoryCardLibraryType = {
  memory: "memory",
  plan: "plan",
  material: "material",
} as const;

export type MemoryCardOperatorType =
  (typeof MemoryCardOperatorType)[keyof typeof MemoryCardOperatorType];

export const MemoryCardOperatorType = {
  user: "user",
  ai: "ai",
} as const;

export interface MemoryCard {
  id: number;
  folderId?: number | null;
  content: string;
  title: string;
  emotionWeight: number;
  emotionLabel: MemoryCardEmotionLabel;
  keywords: string[];
  libraryType: MemoryCardLibraryType;
  operatorType: MemoryCardOperatorType;
  isImportant: boolean;
  priority?: number | null;
  dueDate?: string | null;
  isCompleted?: boolean | null;
  fileType?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateMemoryRequestLibraryType =
  (typeof CreateMemoryRequestLibraryType)[keyof typeof CreateMemoryRequestLibraryType];

export const CreateMemoryRequestLibraryType = {
  memory: "memory",
  plan: "plan",
  material: "material",
} as const;

export interface CreateMemoryRequest {
  content: string;
  title: string;
  folderId?: number | null;
  libraryType: CreateMemoryRequestLibraryType;
  keywords?: string[];
  emotionWeight?: number | null;
  isImportant?: boolean;
  priority?: number | null;
  dueDate?: string | null;
  fileType?: string | null;
}

export interface UpdateMemoryRequest {
  content?: string;
  title?: string;
  folderId?: number | null;
  keywords?: string[];
  emotionWeight?: number | null;
  isImportant?: boolean;
  priority?: number | null;
  dueDate?: string | null;
  isCompleted?: boolean | null;
}

export interface ChatRequest {
  message: string;
  folderId?: number | null;
  useMemories?: boolean;
}

export interface ChatResponse {
  reply: string;
  usedMemories: MemoryCard[];
  newMemoryCreated: boolean;
  messageId?: number;
}

export interface RetrieveMemoriesRequest {
  context: string;
  topK?: number;
  folderId?: number | null;
}

export type RetrieveMemoriesResponseMemoriesItem = {
  memoryId: number;
  content: string;
  title: string;
  relevanceScore: number;
  emotionWeight: number;
};

export interface RetrieveMemoriesResponse {
  memories: RetrieveMemoriesResponseMemoriesItem[];
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  id: number;
  role: ChatMessageRole;
  content: string;
  folderId?: number | null;
  createdAt: string;
}

export type ListMemoriesParams = {
  folderId?: number;
  search?: string;
  libraryType?: ListMemoriesLibraryType;
};

export type ListMemoriesLibraryType =
  (typeof ListMemoriesLibraryType)[keyof typeof ListMemoriesLibraryType];

export const ListMemoriesLibraryType = {
  memory: "memory",
  plan: "plan",
  material: "material",
} as const;

export type GetChatMessagesParams = {
  folderId?: number;
};