# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Memory Management System API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all folders
 */
export const ListFoldersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  parentId: zod.number().nullish(),
  libraryType: zod.enum(["memory", "plan", "material"]),
  createdAt: zod.date(),
  children: zod.array(zod.unknown()).optional(),
  memoryCount: zod.number().optional(),
});
export const ListFoldersResponse = zod.array(ListFoldersResponseItem);

/**
 * @summary Create a folder
 */
export const CreateFolderBody = zod.object({
  name: zod.string(),
  parentId: zod.number().nullish(),
  libraryType: zod.enum(["memory", "plan", "material"]),
});

/**
 * @summary Get a folder
 */
export const GetFolderParams = zod.object({
  folderId: zod.coerce.number(),
});

export const GetFolderResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  parentId: zod.number().nullish(),
  libraryType: zod.enum(["memory", "plan", "material"]),
  createdAt: zod.date(),
  children: zod.array(zod.unknown()).optional(),
  memoryCount: zod.number().optional(),
});

/**
 * @summary Update a folder
 */
export const UpdateFolderParams = zod.object({
  folderId: zod.coerce.number(),
});

export const UpdateFolderBody = zod.object({
  name: zod.string().optional(),
  parentId: zod.number().nullish(),
});

export const UpdateFolderResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  parentId: zod.number().nullish(),
  libraryType: zod.enum(["memory", "plan", "material"]),
  createdAt: zod.date(),
  children: zod.array(zod.unknown()).optional(),
  memoryCount: zod.number().optional(),
});

/**
 * @summary Delete a folder
 */
export const DeleteFolderParams = zod.object({
  folderId: zod.coerce.number(),
});

/**
 * @summary List memories
 */
export const ListMemoriesQueryParams = zod.object({
  folderId: zod.coerce.number().optional(),
  search: zod.coerce.string().optional(),
  libraryType: zod.enum(["memory", "plan", "material"]).optional(),
});

export const ListMemoriesResponseItem = zod.object({
  id: zod.number(),
  folderId: zod.number().nullish(),
  content: zod.string(),
  title: zod.string(),
  emotionWeight: zod.number(),
  emotionLabel: zod.enum([
    "calm",
    "excited",
    "positive",
    "negative",
    "neutral",
  ]),
  keywords: zod.array(zod.string()),
  libraryType: zod.enum(["memory", "plan", "material"]),
  operatorType: zod.enum(["user", "ai"]),
  isImportant: zod.boolean(),
  priority: zod.number().nullish(),
  dueDate: zod.date().nullish(),
  isCompleted: zod.boolean().nullish(),
  fileType: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListMemoriesResponse = zod.array(ListMemoriesResponseItem);

/**
 * @summary Create a memory card
 */
export const CreateMemoryBody = zod.object({
  content: zod.string(),
  title: zod.string(),
  folderId: zod.number().nullish(),
  libraryType: zod.enum(["memory", "plan", "material"]),
  keywords: zod.array(zod.string()).optional(),
  emotionWeight: zod.number().nullish(),
  isImportant: zod.boolean().optional(),
  priority: zod.number().nullish(),
  dueDate: zod.date().nullish(),
  fileType: zod.string().nullish(),
});

/**
 * @summary Get a memory card
 */
export const GetMemoryParams = zod.object({
  memoryId: zod.coerce.number(),
});

export const GetMemoryResponse = zod.object({
  id: zod.number(),
  folderId: zod.number().nullish(),
  content: zod.string(),
  title: zod.string(),
  emotionWeight: zod.number(),
  emotionLabel: zod.enum([
    "calm",
    "excited",
    "positive",
    "negative",
    "neutral",
  ]),
  keywords: zod.array(zod.string()),
  libraryType: zod.enum(["memory", "plan", "material"]),
  operatorType: zod.enum(["user", "ai"]),
  isImportant: zod.boolean(),
  priority: zod.number().nullish(),
  dueDate: zod.date().nullish(),
  isCompleted: zod.boolean().nullish(),
  fileType: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Update a memory card
 */
export const UpdateMemoryParams = zod.object({
  memoryId: zod.coerce.number(),
});

export const UpdateMemoryBody = zod.object({
  content: zod.string().optional(),
  title: zod.string().optional(),
  folderId: zod.number().nullish(),
  keywords: zod.array(zod.string()).optional(),
  emotionWeight: zod.number().nullish(),
  isImportant: zod.boolean().optional(),
  priority: zod.number().nullish(),
  dueDate: zod.date().nullish(),
  isCompleted: zod.boolean().nullish(),
});

export const UpdateMemoryResponse = zod.object({
  id: zod.number(),
  folderId: zod.number().nullish(),
  content: zod.string(),
  title: zod.string(),
  emotionWeight: zod.number(),
  emotionLabel: zod.enum([
    "calm",
    "excited",
    "positive",
    "negative",
    "neutral",
  ]),
  keywords: zod.array(zod.string()),
  libraryType: zod.enum(["memory", "plan", "material"]),
  operatorType: zod.enum(["user", "ai"]),
  isImportant: zod.boolean(),
  priority: zod.number().nullish(),
  dueDate: zod.date().nullish(),
  isCompleted: zod.boolean().nullish(),
  fileType: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete a memory card
 */
export const DeleteMemoryParams = zod.object({
  memoryId: zod.coerce.number(),
});

/**
 * @summary Send a chat message to AI with relevant memories
 */
export const AiChatBody = zod.object({
  message: zod.string(),
  folderId: zod.number().nullish(),
  useMemories: zod.boolean().optional(),
});

export const AiChatResponse = zod.object({
  reply: zod.string(),
  usedMemories: zod.array(
    zod.object({
      id: zod.number(),
      folderId: zod.number().nullish(),
      content: zod.string(),
      title: zod.string(),
      emotionWeight: zod.number(),
      emotionLabel: zod.enum([
        "calm",
        "excited",
        "positive",
        "negative",
        "neutral",
      ]),
      keywords: zod.array(zod.string()),
      libraryType: zod.enum(["memory", "plan", "material"]),
      operatorType: zod.enum(["user", "ai"]),
      isImportant: zod.boolean(),
      priority: zod.number().nullish(),
      dueDate: zod.date().nullish(),
      isCompleted: zod.boolean().nullish(),
      fileType: zod.string().nullish(),
      createdAt: zod.date(),
      updatedAt: zod.date(),
    }),
  ),
  newMemoryCreated: zod.boolean(),
  messageId: zod.number().optional(),
});

/**
 * @summary Retrieve relevant memories for a context
 */
export const retrieveMemoriesBodyTopKDefault = 5;

export const RetrieveMemoriesBody = zod.object({
  context: zod.string(),
  topK: zod.number().default(retrieveMemoriesBodyTopKDefault),
  folderId: zod.number().nullish(),
});

export const RetrieveMemoriesResponse = zod.object({
  memories: zod.array(
    zod.object({
      memoryId: zod.number(),
      content: zod.string(),
      title: zod.string(),
      relevanceScore: zod.number(),
      emotionWeight: zod.number(),
    }),
  ),
});

/**
 * @summary Get chat message history
 */
export const GetChatMessagesQueryParams = zod.object({
  folderId: zod.coerce.number().optional(),
});

export const GetChatMessagesResponseItem = zod.object({
  id: zod.number(),
  role: zod.enum(["user", "assistant"]),
  content: zod.string(),
  folderId: zod.number().nullish(),
  createdAt: zod.date(),
});
export const GetChatMessagesResponse = zod.array(GetChatMessagesResponseItem);