# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Memory Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatRequest {
  message: string;
  folderId?: number | null;
  useMemories?: boolean;
}