# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Memory Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { MemoryCard } from "./memoryCard";

export interface ChatResponse {
  reply: string;
  usedMemories: MemoryCard[];
  newMemoryCreated: boolean;
  messageId?: number;
}