# lib/api-zod/src/generated/types/createFolderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Memory Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateFolderRequestLibraryType } from "./createFolderRequestLibraryType";

export interface CreateFolderRequest {
  name: string;
  parentId?: number | null;
  libraryType: CreateFolderRequestLibraryType;
}