# lib/api-zod/src/generated/types/createMemoryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Memory Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateMemoryRequestLibraryType } from "./createMemoryRequestLibraryType";

export interface CreateMemoryRequest {
  content: string;
  title: string;
  folderId?: number | null;
  libraryType: CreateMemoryRequestLibraryType;
  keywords?: string[];
  emotionWeight?: number | null;
  isImportant?: boolean;
  priority?: number | null;
  dueDate?: Date | null;
  fileType?: string | null;
}