# lib/api-zod/src/generated/types/folder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Memory Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { FolderLibraryType } from "./folderLibraryType";

export interface Folder {
  id: number;
  name: string;
  parentId?: number | null;
  libraryType: FolderLibraryType;
  createdAt: Date;
  children?: Folder[];
  memoryCount?: number;
}