# lib/api-zod/src/generated/types/retrieveMemoriesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Memory Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { RetrieveMemoriesResponseMemoriesItem } from "./retrieveMemoriesResponseMemoriesItem";

export interface RetrieveMemoriesResponse {
  memories: RetrieveMemoriesResponseMemoriesItem[];
}